Use secret-service 1.5.0. 

Closes #2 .

+++ src/main/java/org/cryptomator/linux/keychain/SecretServiceKeychainAccess.java
@@ -5,6 +5,7 @@ import org.cryptomator.integrations.keychain.KeychainAccessProvider;
 import org.freedesktop.secret.simple.SimpleCollection;
 
 import java.io.IOException;
+import java.security.AccessControlException;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
@@ -15,24 +16,19 @@ public class SecretServiceKeychainAccess implements KeychainAccessProvider {
 
 	@Override
 	public boolean isSupported() {
-		try (@SuppressWarnings("unused") SimpleCollection keyring = new SimpleCollection()) {
-			// seems like we're able to access the keyring.
-			return true;
-		} catch (IOException | ExceptionInInitializerError | RuntimeException e) {
-			return false;
-		}
+		return SimpleCollection.isAvailable();
 	}
 
 	@Override
 	public void storePassphrase(String key, CharSequence passphrase) throws KeychainAccessException {
 		try (SimpleCollection keyring = new SimpleCollection()) {
 			List<String> list = keyring.getItems(createAttributes(key));
-			if (list == null) {
+			if (list == null || list.isEmpty()) {
 				keyring.createItem(LABEL_FOR_SECRET_IN_KEYRING, passphrase, createAttributes(key));
 			} else {
 				changePassphrase(key, passphrase);
 			}
-		} catch (IOException e) {
+		} catch (IOException | AccessControlException e) {
 			throw new KeychainAccessException("Storing password failed.", e);
 		}
 	}
@@ -41,12 +37,12 @@ public class SecretServiceKeychainAccess implements KeychainAccessProvider {
 	public char[] loadPassphrase(String key) throws KeychainAccessException {
 		try (SimpleCollection keyring = new SimpleCollection()) {
 			List<String> list = keyring.getItems(createAttributes(key));
-			if (list != null) {
+			if (list != null && !list.isEmpty()) {
 				return keyring.getSecret(list.get(0));
 			} else {
 				return null;
 			}
-		} catch (IOException e) {
+		} catch (IOException | AccessControlException e) {
 			throw new KeychainAccessException("Loading password failed.", e);
 		}
 	}
@@ -55,10 +51,10 @@ public class SecretServiceKeychainAccess implements KeychainAccessProvider {
 	public void deletePassphrase(String key) throws KeychainAccessException {
 		try (SimpleCollection keyring = new SimpleCollection()) {
 			List<String> list = keyring.getItems(createAttributes(key));
-			if (list != null) {
+			if (list != null && !list.isEmpty()) {
 				keyring.deleteItem(list.get(0));
 			}
-		} catch (IOException e) {
+		} catch (IOException | AccessControlException e) {
 			throw new KeychainAccessException("Deleting password failed.", e);
 		}
 	}
@@ -67,10 +63,10 @@ public class SecretServiceKeychainAccess implements KeychainAccessProvider {
 	public void changePassphrase(String key, CharSequence passphrase) throws KeychainAccessException {
 		try (SimpleCollection keyring = new SimpleCollection()) {
 			List<String> list = keyring.getItems(createAttributes(key));
-			if (list != null) {
+			if (list != null && !list.isEmpty()) {
 				keyring.updateItem(list.get(0), LABEL_FOR_SECRET_IN_KEYRING, passphrase, createAttributes(key));
 			}
-		} catch (IOException e) {
+		} catch (IOException | AccessControlException e) {
 			throw new KeychainAccessException("Changing password failed.", e);
 		}
 	}
@@ -78,4 +74,5 @@ public class SecretServiceKeychainAccess implements KeychainAccessProvider {
 	private Map<String, String> createAttributes(String key) {
 		return Map.of("Vault", key);
 	}
+
 }
